+++ cx-portal-shared-components/src/components/basic/Table/components/Toolbar/index.tsx
@@ -29,6 +29,7 @@ import FilterIcon from '@mui/icons-material/FilterAltOutlined'
 import ClearIcon from '@mui/icons-material/Clear'
 import { Checkbox } from '../../../Checkbox'
 import { getSelectedFilterUpdate } from './helper'
+import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline'
 
 interface FilterValue {
   value: string
@@ -87,8 +88,10 @@ export const Toolbar = ({
   onClearSearch,
 }: ToolbarProps) => {
   const { spacing } = useTheme()
+  const isSearchText = searchExpr && searchExpr !== ''
+  const isSearchData = searchInputData ? searchInputData.open : false
   const [openSearch, setOpenSearch] = useState<boolean>(
-    searchInputData ? searchInputData.open : false
+    isSearchText ? isSearchText : isSearchData
   )
   const [openFilter, setOpenFilter] = useState<boolean>(false)
   const [searchInput, setSearchInput] = useState<string>(
@@ -181,7 +184,12 @@ export const Toolbar = ({
             </Box>
           </Typography>
           {buttonLabel && onButtonClick && (
-            <Button size="small" onClick={onButtonClick} sx={{ marginLeft: 3 }}>
+            <Button
+              startIcon={<AddCircleOutlineIcon />}
+              size="small"
+              onClick={onButtonClick}
+              sx={{ marginLeft: 15 }}
+            >
               {buttonLabel}
             </Button>
           )}
@@ -195,8 +203,14 @@ export const Toolbar = ({
               value={searchInput}
               onChange={onSearchInputChange}
               onKeyPress={onSearchInputKeyPress}
-              onBlur={() => setOpenSearch(false)}
+              onBlur={() => {}}
               placeholder={searchPlaceholder}
+              sx={{
+                '.MuiInputBase-input': {
+                  padding: '10px',
+                  width: '250px',
+                },
+              }}
             />
           ) : (
             onSearch && (