+++ cx-portal/src/components/pages/Admin/components/RegistrationRequests/components/RequestList/index.tsx
@@ -37,6 +37,8 @@ export const RequestList = ({
   isLoading,
   onTableCellClick,
   loaded,
+  handleDownloadDocument,
+  searchExpr,
 }: {
   fetchHook: (paginArgs: PaginFetchArgs) => any
   fetchHookArgs?: any
@@ -46,6 +48,8 @@ export const RequestList = ({
   isLoading: boolean
   onTableCellClick: (params: GridCellParams) => void
   loaded: number
+  handleDownloadDocument: (documentId: string, documentType: string) => void
+  searchExpr?: string
 }) => {
   const { t } = useTranslation()
   const dispatch = useDispatch()
@@ -56,7 +60,8 @@ export const RequestList = ({
     useTranslation,
     onApproveClick,
     onDeclineClick,
-    isLoading
+    isLoading,
+    handleDownloadDocument
   )
 
   useEffect(() => {
@@ -72,6 +77,7 @@ export const RequestList = ({
   return (
     <section id="identity-management-id">
       <PageLoadingTable<ApplicationRequest>
+        searchExpr={searchExpr}
         rowHeight={80}
         onCellClick={onTableCellClick}
         toolbarVariant={'ultimate'}